chore: move binary decode to socket.rs
